requirements.txt
streamlit
pickle
mysql.connector
requests
pinecone
PyMuPDF
google-generativeai
os
